Collision Manager should be finished
