+++ src/dinojump/manager/CollisionManager.java
@@ -1,0 +1,62 @@
+package dinojump.manager;
+
+import dinojump.Renderer;
+import dinojump.util.Avatar;
+import dinojump.util.Position;
+import list.List;
+
+public class CollisionManager
+{
+    private static CollisionManager instance;
+    private Avatar avatar;
+    private List platforms;
+
+    private CollisionManager()
+    {
+
+    }
+
+    public static CollisionManager getInstance()
+    {
+        if (instance == null)
+        {
+            return new CollisionManager();
+        }
+        else
+        {
+            return instance;
+        }
+
+    }
+
+    public void setPlatforms(List newPlatforms)
+    {
+        platforms = newPlatforms;
+    }
+
+    public void checkForCollision()
+    {
+        Position avtPos = avatar.getPosition();
+        platforms = PlatformManager.getInstance().getPlatforms();
+        int avtVel = avatar.getYVelocity();
+
+        if (avtVel < 0)
+        {
+            for (int i = 0; i < platforms.size(); i++)
+            {
+                if (avtPos.getX() + Renderer.getInstance().getPlatformWidth() < platforms.get(i).getPosition().getX() && avtPos.getX() + Renderer.getInstance().getAvatarDimensions() > platforms.get(i).getPosition().getX())
+                {
+                    if (avtPos.getY() > platforms.get(i).getPosition().getY() && avtPos.getY() + Renderer.getInstance().getPlatformHeight() < platforms.get(i).getPosition().getY())
+                    {
+                        Avatar.getInstance().setYAcceleration(1);
+                    }
+                }
+
+            }
+        }
+
+    }
+}
+
+
+
