+++ src/dinojump/manager/PlatformManager.java
@@ -73,4 +73,9 @@ public class PlatformManager
     {
         Renderer.getInstance().renderPlatforms(platforms);
     }
+
+    public List getPlatforms()
+    {
+        return platforms;
+    }
 }
